+++ src/components/mount/infoMount.vue
@@ -22,11 +22,27 @@
       :enabledText="$t('components.mount.info.slewing')"
       :disabledText="$t('components.mount.info.notSlewing')"
     />
+    <StatusString
+      :isEnabled="store.mountInfo.RightAscensionString"
+      :Name="$t('components.mount.info.rightAscension')"
+      :Value="store.mountInfo.RightAscensionString || ''"
+    />
+    <StatusString
+      :isEnabled="store.mountInfo.DeclinationString"
+      :Name="$t('components.mount.info.declination')"
+      :Value="store.mountInfo.DeclinationString || ''"
+    />
+    <StatusString
+      :isEnabled="store.mountInfo.TimeToMeridianFlipString"
+      :Name="$t('components.mount.info.timeToMeridianFlip')"
+      :Value="store.mountInfo.TimeToMeridianFlipString || ''"
+    />
   </div>
 </template>
 
 <script setup>
 import StatusBool from '@/components/helpers/StatusBool.vue';
+import StatusString from '@/components/helpers/StatusString.vue';
 import { apiStore } from '@/store/store';
 const store = apiStore();